cleanup alice swap_state.go, move funcs
